incomplete Hosts/final/turin_run.sh

+++ final/turin_run.sh
@@ -68,6 +68,9 @@ if [[ $current_state == $state ]]; then
                 ./result.sh > /var/lib/lkp-files/results/lkpvms.txt
 
 	elif [[ $test_bed == $case4 ]]; then
+		sudo touch /var/lib/lkp-files/test_progress.txt
+		test_progress=/var/lib/lkp-files/test_progress.txt
+		lkp_progress=/var/lib/lkp-files/progress.txt
 		if [ -d /var/lib/lkp-files/results/all ]; then
 			rm -rf /var/lib/lkp-files/results/all
 		fi
@@ -76,14 +79,29 @@ if [[ $current_state == $state ]]; then
                 sudo rm -rf /lkp/result/ebizzy/*
                 sudo rm -rf /lkp/result/unixbench/*
 		echo "Run all test beds"
-		$loc/lkp-tests/lkp.sh
-
-		echo "Archiving results"
-                cd /lkp/result/
-                ./result.sh > /var/lib/lkp-files/results/all/novms.txt
-		sudo rm -rf /lkp/result/hackbench/*
-                sudo rm -rf /lkp/result/ebizzy/*
-                sudo rm -rf /lkp/result/unixbench/*
+		var1=$(cat test_progress)
+		var2="novms"
+		var3="vms"
+		var4="lkpvms"
+		if [[ $var1 == $var2 ]]; then
+			$loc/lkp-tests/lkp.sh
+			if [[ -f $lkp_progress ]]; then
+				echo ""
+			else
+				#resume archiving results
+				echo "Archiving results"
+			echo "vms" > $test_progress
+		fi
+		if [[ -f $lkp_progress ]]; then
+			echo ""
+		else
+			echo "Archiving results"
+			cd /lkp/result/
+			./result.sh > /var/lib/lkp-files/results/all/novms.txt
+			sudo rm -rf /lkp/result/hackbench/*
+			sudo rm -rf /lkp/result/ebizzy/*
+			sudo rm -rf /lkp/result/unixbench/*
+		fi
 
 		if [[ $stress_type == $case1 ]]; then
 			echo "stres-ng"
